// src/api-utils/generated-client/api.ts
export interface GetRepaymentDto {
    /**
     * 
     * @type {number}
     * @memberof GetRepaymentDto
     */
    'bank_account_number': number;
    /**
     * 
     * @type {string}
     * @memberof GetRepaymentDto
     */
    'withdrawl_bank_name': string;
    /**
     * 
     * @type {number}
     * @memberof GetRepaymentDto
     */
    'expected_total_payback': number;
    /**
     * 
     * @type {object}
     * @memberof GetRepaymentDto
     */
    'full_remaining_balance_incl_interest': object;
    /**
     * 
     * @type {string}
     * @memberof GetRepaymentDto
     */
    'funded_date': string;
    /**
     * 
     * @type {number}
     * @memberof GetRepaymentDto
     */
    'funding_amount': number;
    /**
     * 
     * @type {string}
     * @memberof GetRepaymentDto
     */
    'next_hit_date': string;
    /**
     * 
     * @type {number}
     * @memberof GetRepaymentDto
     */
    'next_interest_amount': number;
    /**
     * 
     * @type {string}
     * @memberof GetRepaymentDto
     */
    'next_interest_date': string;
    /**
     * 
     * @type {string}
     * @memberof GetRepaymentDto
     */
    'payoff': string;
    /**
     * 
     * @type {string}
     * @memberof GetRepaymentDto
     */
    'payroll_record': string;
    /**
     * 
     * @type {object}
     * @memberof GetRepaymentDto
     */
    'payroll_record__r': object;
    /**
     * 
     * @type {string}
     * @memberof GetRepaymentDto
     */
    'status': string;
    /**
     * 
     * @type {string}
     * @memberof GetRepaymentDto
     */
    'salesforce_id': string;
    /**
     * 
     * @type {number}
     * @memberof GetRepaymentDto
     */
    'total_amount_settled': number;
    /**
     * 
     * @type {number}
     * @memberof GetRepaymentDto
     */
    'total_outstanding_amount': number;
    /**
     * 
     * @type {number}
     * @memberof GetRepaymentDto
     */
    'total_outstanding_interest': number;
    /**
     * 
     * @type {number}
     * @memberof GetRepaymentDto
     */
    'total_outstanding_principal': number;
    /**
     * 
     * @type {string}
     * @memberof GetRepaymentDto
     */
    'number_of_pay_periods': string;
    /**
     * 
     * @type {number}
     * @memberof GetRepaymentDto
     */
    'total_credits_settled': number;
    /**
     * 
     * @type {string}
     * @memberof GetRepaymentDto
     */
    'merchant_payroll_company': string;
    /**
     * 
     * @type {string}
     * @memberof GetRepaymentDto
     */
    'uuid': string;
}
/**
 * 
 * @export
 * @interface LoanPayment
 */
export interface LoanPayment {
    /**
     * 
     * @type {string}
     * @memberof LoanPayment
     */
    'repayment': string;
    /**
     * 
     * @type {string}
     * @memberof LoanPayment
     */
    'repayment_r.uuid__c': string;
    /**
     * 
     * @type {string}
     * @memberof LoanPayment
     */
    'repayment_account_name': string;
    /**
     * 
     * @type {string}
     * @memberof LoanPayment
     */
    'bank_name_from_repayment': string;
    /**
     * 
     * @type {string}
     * @memberof LoanPayment
     */
    'bank_account_from_repayment': string;
    /**
     * 
     * @type {string}
     * @memberof LoanPayment
     */
    'repayment_bank_account_number': string;
    /**
     * 
     * @type {string}
     * @memberof LoanPayment
     */
    'repayment_bank_name': string;
    /**
     * 
     * @type {number}
     * @memberof LoanPayment
     */
    'repayment_funding_amount': number;
    /**
     * 
     * @type {number}
     * @memberof LoanPayment
     */
    'running_balance': number;
    /**
     * 
     * @type {number}
     * @memberof LoanPayment
     */
    'running_fee_balance': number;
    /**
     * 
     * @type {number}
     * @memberof LoanPayment
     */
    'running_principal': number;
    /**
     * 
     * @type {string}
     * @memberof LoanPayment
     */
    'settlement_date': string;
    /**
     * 
     * @type {string}
     * @memberof LoanPayment
     */
    'effective_date': string;
    /**
     * 
     * @type {string}
     * @memberof LoanPayment
     */
    'status': string;
    /**
     * 
     * @type {string}
     * @memberof LoanPayment
     */
    'status_amount': string;
    /**
     * 
     * @type {string}
     * @memberof LoanPayment
     */
    'sub_type': string;
    /**
     * 
     * @type {number}
     * @memberof LoanPayment
     */
    'transaction_amount': number;
    /**
     * 
     * @type {string}
     * @memberof LoanPayment
     */
    'transaction_type': string;
    /**
     * 
     * @type {string}
     * @memberof LoanPayment
     */
    'uuid': string;
}
// ...
export interface NewLoanDebitPayment {
    /**
     * 
     * @type {number}
     * @memberof NewLoanDebitPayment
     */
    'amount': number;
    /**
     * 
     * @type {string}
     * @memberof NewLoanDebitPayment
     */
    'effectiveDate': string;
    /**
     * 
     * @type {string}
     * @memberof NewLoanDebitPayment
     */
    'repaymentId': string;
}